Changes for showing restaurants in UI
