+++ src/main/resources/application.properties
@@ -1,2 +1,3 @@
 client_id=H4BNECr10DAbYLJZ-6ppBg
+yelp_api_url=https://api.yelp.com/v3/businesses/search
 api_key=-3viEYwHOmenfpRiVWvs1PWXFxNVe77RQxnFUIAjyPbjePPmt2FmCapNDsEpkZr2j1I-UIPr8177APs5N64pg_9kgvuQP0-getDJPpoJaTO4mPkRjGf6-x38zLd3ZXYx
